needed to do some more refactoring to uplift the review-state as I want to now implement the delete and edit-logic
